Now AI also can create units

+++ Assets/Scripts/Buildings/Building.cs
@@ -9,7 +9,7 @@ public class Building : MonoBehaviour
     [SerializeField] protected Vector2 size;
     [SerializeField] protected int cost;
     [SerializeField] protected GameObject _explosion_prefab;
-    [SerializeField] protected int _maintenance_cost;
+    [SerializeField] protected float _maintenance_cost;
     public bool IsEnemy { get; set; }
 
     protected void Start()
@@ -46,6 +46,6 @@ public class Building : MonoBehaviour
 
     public int GetCost() => cost;
 
-    public int MaintenanceCost => _maintenance_cost;
+    public float MaintenanceCost => _maintenance_cost;
  
 }
